move feature toggles out of the inspector
